requirements.txt
python-gnupg~=0.4.7